Function mpir::mpz::mpz_get_str

pub unsafe extern "C" fn mpz_get_str(
    s: *mut c_char, 
    base: c_int, 
    op: mpz_srcptr
) -> *mut c_char

Convert op to a string of digits in base base. The base may vary from 2 to 36 or from −2 to −36.

For base in the range 2..36, digits and lower-case letters are used; for −2..−36, digits and upper-case letters are used; for 37..62, digits, upper-case letters, and lower-case letters (in that significance order) are used.

If str is NULL, the result string is allocated using the current allocation function (see Chapter 14 [Custom Allocation], page 106). The block will be strlen(str)+1 bytes, that being exactly enough for the string and null-terminator.

If str is not NULL, it should point to a block of storage large enough for the result, that being mpz_sizeinbase (op, base) + 2. The two extra bytes are for a possible minus sign, and the null-terminator.

A pointer to the result string is returned, being either the allocated block, or the given str.